feat: Add Banner component and export hooks
