+++ src/components/Banner/index.ts
@@ -1,0 +1,2 @@
+export { Banner } from "./Banner"
+export type { BannerProps, IBannerFields } from "./Banner.types"
